# Expectation maximisation speedup option