--- a/splink/em_training_session.py
+++ b/splink/em_training_session.py
@@ -40,7 +40,7 @@ class EMTrainingSession:
         fix_probability_two_random_records_match: bool = False,
         comparisons_to_deactivate: list[Comparison] = None,
         comparison_levels_to_reverse_blocking_rule: list[ComparisonLevel] = None,
-        estimate_without_term_frequencies: bool = True,
+        estimate_without_term_frequencies: bool = False,
     ):
         logger.info("\n----- Starting EM training session -----\n")