--- a/splink/expectation_maximisation.py
+++ b/splink/expectation_maximisation.py
@@ -9,7 +9,7 @@ import pandas as pd
 from .comparison_level import ComparisonLevel
 from .constants import LEVEL_NOT_OBSERVED_TEXT
 from .m_u_records_to_parameters import m_u_records_to_lookup_dict
-from .predict import predict_from_comparison_vectors_sqls
+from .predict import predict_from_agreement_pattern_counts_sqls, predict_from_comparison_vectors_sqls
 from .settings import Settings
 from .splink_dataframe import SplinkDataFrame
 
@@ -21,14 +21,34 @@ if TYPE_CHECKING:
 logger = logging.getLogger(__name__)
 
 
+def count_agreement_patterns_sql(settings_obj: Settings):
+
+    gamma_cols = [cc._gamma_column_name for cc in settings_obj.comparisons]
+    gamma_cols_expr = ','.join(gamma_cols)
+
+    sql = f"""
+    select
+    {gamma_cols_expr},
+    count(*) as agreement_pattern_count
+    from __splink__df_comparison_vectors
+    group by {gamma_cols_expr}
+    """
+
+    return sql
+
+
 def compute_new_parameters_sql(settings_obj: Settings):
     """compute m and u counts from the results of predict"""
+    if getattr(settings_obj, "_predict_on_agreement_counts", False):
+        agreement_pattern_count = "agreement_pattern_count"
+    else:
+        agreement_pattern_count = "1"
 
     sql_template = """
     select
     {gamma_column} as comparison_vector_value,
-    sum(match_probability) as m_count,
-    sum(1-match_probability) as u_count,
+    sum(match_probability * {agreement_pattern_count}) as m_count,
+    sum((1-match_probability) * {agreement_pattern_count}) as u_count,
     '{output_column_name}' as output_column_name
     from __splink__df_predict
     group by {gamma_column}
@@ -37,15 +57,16 @@ def compute_new_parameters_sql(settings_obj: Settings):
         sql_template.format(
             gamma_column=cc._gamma_column_name,
             output_column_name=cc._output_column_name,
+            agreement_pattern_count=agreement_pattern_count,
         )
         for cc in settings_obj.comparisons
     ]
 
     # Probability of two random records matching
-    sql = """
+    sql = f"""
     select 0 as comparison_vector_value,
-           avg(match_probability) as m_count,
-           avg(1-match_probability) as u_count,
+           sum(match_probability * {agreement_pattern_count})/sum({agreement_pattern_count}) as m_count,
+           sum((1-match_probability) * {agreement_pattern_count})/sum({agreement_pattern_count}) as u_count,
            '_probability_two_random_records_match' as output_column_name
     from __splink__df_predict
     """
@@ -196,20 +217,37 @@ def expectation_maximisation(
     em_convergece = settings_obj._em_convergence
     logger.info("")  # newline
 
+    if settings_obj._predict_on_agreement_counts:
+        sql = count_agreement_patterns_sql(settings_obj)
+        linker._enqueue_sql(sql, "__splink__agreement_pattern_counts")
+        agreement_pattern_counts = linker._execute_sql_pipeline(
+            [df_comparison_vector_values]
+        )
+
     for i in range(1, max_iterations + 1):
         start_time = time.time()
 
         # Expectation step
-        sqls = predict_from_comparison_vectors_sqls(
-            settings_obj,
-            sql_infinity_expression=linker._infinity_expression,
-        )
+        if settings_obj._predict_on_agreement_counts:
+            sqls = predict_from_agreement_pattern_counts_sqls(
+                settings_obj,
+                sql_infinity_expression=linker._infinity_expression,
+            )
+        else:
+            sqls = predict_from_comparison_vectors_sqls(
+                settings_obj, 
+                sql_infinity_expression=linker._infinity_expression,
+            )
+
         for sql in sqls:
             linker._enqueue_sql(sql["sql"], sql["output_table_name"])
-
+        
         sql = compute_new_parameters_sql(settings_obj)
         linker._enqueue_sql(sql, "__splink__m_u_counts")
-        df_params = linker._execute_sql_pipeline([df_comparison_vector_values])
+        if settings_obj._predict_on_agreement_counts:
+            df_params = linker._execute_sql_pipeline([agreement_pattern_counts])
+        else: 
+            df_params = linker._execute_sql_pipeline([df_comparison_vector_values])
         param_records = df_params.as_pandas_dataframe()
         param_records = compute_proportions_for_new_parameters(param_records)