--- a/splink/linker.py
+++ b/splink/linker.py
@@ -1485,10 +1485,10 @@ class Linker:
         blocking_rule: str,
         comparisons_to_deactivate: list[str | Comparison] = None,
         comparison_levels_to_reverse_blocking_rule: list[ComparisonLevel] = None,
+        estimate_without_term_frequencies: bool = False,
         fix_probability_two_random_records_match: bool = False,
         fix_m_probabilities=False,
         fix_u_probabilities=True,
-        predict_on_agreement_counts=True,
         populate_probability_two_random_records_match_from_trained_values=False,
     ) -> EMTrainingSession:
         """Estimate the parameters of the linkage model using expectation maximisation.
@@ -1556,6 +1556,10 @@ class Linker:
                 specified in the blocking rule. If provided, this argument will overrule
                 this default behaviour. The user must provide a list of ComparisonLevel
                 objects.  Defaults to None.
+            estimate_without_term_frequencies (bool, optional): If True, the iterations
+                of the EM algorithm ignore any term frequency adjustments and only
+                depend on the comparison vectors. This allows the EM algorithm to run
+                much faster, but the estimation of the parameters will change slightly.
             fix_probability_two_random_records_match (bool, optional): If True, do not
                 update the probability two random records match after each iteration.
                 Defaults to False.
@@ -1610,7 +1614,7 @@ class Linker:
             fix_probability_two_random_records_match=fix_probability_two_random_records_match,  # noqa 501
             comparisons_to_deactivate=comparisons_to_deactivate,
             comparison_levels_to_reverse_blocking_rule=comparison_levels_to_reverse_blocking_rule,  # noqa 501
-            predict_on_agreement_counts=predict_on_agreement_counts,
+            estimate_without_term_frequencies=estimate_without_term_frequencies,
         )
 
         em_training_session._train()