### Type of PR

FEAT

### Is your Pull Request linked to an existing Issue or Pull Request?
Relates to #1363 



### Give a brief description for the solution you have provided
<!--
  Provide a clear and concise description of what you want to happen.
-->



### PR Checklist

- [ ] Added documentation for changes
- [ ] Added feature to example notebooks at tutorials in [splink_demos](https://github.com/moj-analytical-services/splink_demos) (if appropriate)
- [x] Added tests (if appropriate)
- [x] Made changes based off the latest version of Splink
- [x] Run the [linter](https://moj-analytical-services.github.io/splink/dev_guides/changing_splink/lint.html)