@aymonwuolanne 2023-06-27
> This is great - thanks! The core logic all looks good to me.
> 
> There's a minor adjustment i'd like to make around the settings object, just for consistency with the rest of the codebase - are you happy for me to add commits on top?

Absolutely, feel free.