@RobinL 2023-06-28
I tried it out on our 1m row test dataset from Wikidata:

Blocking run 1:
New method: `Iteration time: 0.045393943786621094 seconds`
Old method: `Iteration time: 3.0039303302764893 seconds`
🎉 60x speedup

Blocking run 2, with a looser rule:
New method: `Iteration time: 0.12349581718444824 seconds`
Old method : `Iteration time: 104 seconds`
🎉 1000x speedup